requirements.txt
openai
requests
Pillow
flask
flask-cors
python-dotenv